+++ indexer/src/test/java/org/pdxfinder/commandline/FinderExporterTest.java
@@ -4,11 +4,12 @@ import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.Spy;
 import org.pdxfinder.BaseTest;
+import org.pdxfinder.dataexport.UniversalDataExtractor;
 import org.pdxfinder.services.DataImportService;
-import org.pdxfinder.services.UtilityService;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.boot.test.mock.mockito.SpyBean;
 
 import java.io.File;
 import java.io.IOException;
@@ -23,33 +24,47 @@ public class FinderExporterTest extends BaseTest {
     public TemporaryFolder folder = new TemporaryFolder();
     private File tempFile;
 
-    @MockBean
-    private UtilityService utilityService;
-    @MockBean
+    @Mock
     private DataImportService dataImportService;
+    @Mock
+    private UniversalDataExtractor universalDataExtractor;
 
-    @SpyBean
-    FinderExporter finderExporter;
+    @Spy
+    @InjectMocks
+    private FinderExporter finderExporter;
 
     @Before public void init() throws IOException {
         tempFile = folder.newFile();
     }
 
     @Test
-    public void Given_loadAllisTrue_When_runIsCalled_Then_CallExportAll() throws IOException {
+    public void given_loadAllisTrue_When_runIsCalled_Then_CallExportAll() throws IOException {
         finderExporter.setDefaultDirectory(tempFile.getAbsolutePath());
-        finderExporter.run(null, null, true);
-        verify(finderExporter).exportAllGroups(any(File.class)) ;
+        finderExporter.run(null, null, true,false );
+        verify(finderExporter).exportAllGroups(any(File.class), false) ;
     }
 
     @Test
-    public void Given_provider_when_runIsCalled_Then_CallExport() throws IOException {
+    public void given_provider_when_runIsCalled_Then_CallExport() throws IOException {
         finderExporter.setDefaultDirectory(tempFile.getAbsolutePath());
-        finderExporter.run(null, "test", false);
+        finderExporter.run(null, "test", false,false);
         verify(finderExporter).export(
                 eq(tempFile.getAbsoluteFile()),
-                eq("test")
+                eq("test"),
+                eq(false)
         );
+    }
+
+    @Test
+    public void given_WithSingleProviderAndharmonizedIsTrue_when_runisCalled_Then_CallExport() throws IOException {
+        finderExporter.setDefaultDirectory(tempFile.getAbsolutePath());
+        finderExporter.run(null, "test", false, true);
+        verify(finderExporter).export(
+                eq(tempFile.getAbsoluteFile()),
+                eq("test"),
+                eq(true)
+        );
+
     }
 
 }
