+++ web/src/main/resources/templates/details.html
@@ -868,7 +868,7 @@
                                                     </td>
                                                     <td>
                                                         <a th:text="${dataRow.rawDataLabel}"
-                                                           style="..."
+                                                           style="color: #06369d; text-decoration: none;"
                                                            th:onclick="'window.open(\'https://www.ebi.ac.uk/ena/data/view/'+${dataRow.rawDataLabel}+'\')'">
                                                             ...
                                                         </a>
